Add files via upload
The Frog Leap Problem is a classic puzzle that involves a frog trying to hop across a river by jumping on stones. The objective is to find the minimum number of jumps required for the frog to reach the other side of the river, given certain constraints such as the size of the river, the positions of the stones, and the maximum distance the frog can jump.

This Python project aims to implement a program that can solve the Frog Leap Problem efficiently. The program will take inputs such as the size of the river, the positions of the stones, and the maximum jump distance of the frog, and then output the minimum number of jumps required.